refactor(ui): make OAuth routes statically /dex and drop provider plumbing
The `[provider]` dynamic segment was guarding a single value ("dex");
fold it into the URL itself so the constraint becomes structural rather
than a runtime check.

- Move ui/src/app/api/oauth/start/[provider]/route.ts and
  callback/[provider]/route.ts to .../start/dex/ and .../callback/dex/.
  Drop the params arg, the provider extraction, and the
  `if (provider === "dex")` guard from each handler.
- The callback route no longer writes `provider=dex` into the redirect
  URL back to /login. Drop the now-redundant `dex:` prefix on the
  oauth-state cookie (was `${provider}:${state}`, now just `state`).
- ui/src/app/login/page.tsx: stop reading `provider` from search params;
  signIn("backend-code", { code }) — no provider field.
- ui/src/auth.config.ts: drop the `provider` credential from the
  backend-code Credentials provider; rename oauthCallbackRedirectUri →
  dexCallbackRedirectUri (no arg, hardcodes /api/oauth/callback/dex);
  the POST to /auth/oauth/exchange hardcodes provider: "dex" in the body.
- Tests: drop `&provider=dex` from URL fixtures and `provider: 'dex'`
  from signIn assertions in the two affected cases.

The backend API contract (/auth/oauth-url?provider=…,
OAuthExchangeRequest.provider) is intentionally untouched — the UI just
always sends "dex" now. Dropping `provider` server-side is a separate
ticket.
