+++ ui/src/app/api/oauth/callback/dex/route.ts
@@ -2,11 +2,7 @@ import {NextRequest, NextResponse} from "next/server";
 import {tryResolvePublicOrigin} from "@/lib/server/resolve-public-origin-response";
 import {sanitizeCallbackUrl} from "@/lib/utils/url";
 
-export async function GET(
-  request: NextRequest,
-  { params }: { params: Promise<{ provider: string }> }
-) {
-  const {provider} = await params;
+export async function GET(request: NextRequest) {
   const searchParams = request.nextUrl.searchParams;
   const code = searchParams.get("code");
   const error = searchParams.get("error");
@@ -24,10 +20,7 @@ export async function GET(
   const rawCallbackUrl = request.cookies.get("oauth-callback-url")?.value || "/";
   loginUrl.searchParams.set("callbackUrl", sanitizeCallbackUrl(rawCallbackUrl));
 
-  const stateCookie = request.cookies.get("oauth-state")?.value;
-  const expectedState = stateCookie?.startsWith(`${provider}:`)
-    ? stateCookie.slice(provider.length + 1)
-    : undefined;
+  const expectedState = request.cookies.get("oauth-state")?.value;
   if (!expectedState || !returnedState || expectedState !== returnedState) {
     console.error("OAuth state mismatch — possible CSRF or provider confusion");
     loginUrl.searchParams.set("error", "authentication_failed");
@@ -37,18 +30,14 @@ export async function GET(
     return redirectResponse;
   }
 
-  // Add provider and code/error parameters BEFORE creating the redirect response
-  if (provider === "dex") {
-    loginUrl.searchParams.set("provider", provider);
-    if (code) {
-      loginUrl.searchParams.set("code", code);
-    } else if (error) {
-      const KNOWN_ERRORS = ["access_denied", "user_not_allowed", "server_error", "temporarily_unavailable"];
-      const safeError = KNOWN_ERRORS.includes(error) ? error : "authentication_failed";
-      loginUrl.searchParams.set("error", safeError);
-    } else {
-      loginUrl.searchParams.set("error", "No authorization code received");
-    }
+  if (code) {
+    loginUrl.searchParams.set("code", code);
+  } else if (error) {
+    const KNOWN_ERRORS = ["access_denied", "user_not_allowed", "server_error", "temporarily_unavailable"];
+    const safeError = KNOWN_ERRORS.includes(error) ? error : "authentication_failed";
+    loginUrl.searchParams.set("error", safeError);
+  } else {
+    loginUrl.searchParams.set("error", "No authorization code received");
   }
 
   // Create redirect response with the complete URL including all parameters
@@ -59,4 +48,3 @@ export async function GET(
 
   return redirectResponse;
 }
-
